feat(frontend): Location data updated

See merge request se/elektrobit/elektrobit!38

+++ backend/src/main/kotlin/de/uulm/automotive/cds/entities/LocationData.kt
@@ -5,5 +5,5 @@ import javax.persistence.GeneratedValue
 import javax.persistence.Id
 
 @Entity
-class LocationData (@Id @GeneratedValue var id: Long?, var lat: Long, var lng: Long, var radius: Int) {
+class LocationData (@Id @GeneratedValue var id: Long?, var lat: Double, var lng: Double, var radius: Int) {
 }
